moved util functions to UserAgent class
